two_finger_gripper class to abstract gripper from base_mp + cleanup comments/class fields
