--- mani_skill/examples/motionplanning/base_motionplanner/gripper_pose_visual_builders.py
@@ -1,149 +1,0 @@
-import numpy as np
-import sapien
-from transforms3d import quaternions
-from mani_skill.envs.scene import ManiSkillScene
-
-
-def build_robotiq_gripper_grasp_pose_visual(scene: ManiSkillScene):
-    builder = scene.create_actor_builder()
-    grasp_pose_visual_width = 0.01
-    grasp_width = 0.05
-
-    builder.add_sphere_visual(
-        pose=sapien.Pose(p=[0, 0, 0.0]),
-        radius=grasp_pose_visual_width,
-        material=sapien.render.RenderMaterial(base_color=[0.3, 0.4, 0.8, 0.7])
-    )
-
-    builder.add_box_visual(
-        pose=sapien.Pose(p=[0, 0, -0.08]),
-        half_size=[grasp_pose_visual_width, grasp_pose_visual_width, 0.02],
-        material=sapien.render.RenderMaterial(base_color=[0, 1, 0, 0.7]),
-    )
-    builder.add_box_visual(
-        pose=sapien.Pose(p=[0, 0, -0.05]),
-        half_size=[grasp_pose_visual_width, grasp_width, grasp_pose_visual_width],
-        material=sapien.render.RenderMaterial(base_color=[0, 1, 0, 0.7]),
-    )
-    builder.add_box_visual(
-        pose=sapien.Pose(
-            p=[
-                0.03 - grasp_pose_visual_width * 3,
-                grasp_width + grasp_pose_visual_width,
-                0.03 - 0.05,
-            ],
-            q=quaternions.axangle2quat(np.array([0, 1, 0]), theta=np.pi / 2),
-        ),
-        half_size=[0.04, grasp_pose_visual_width, grasp_pose_visual_width],
-        material=sapien.render.RenderMaterial(base_color=[0, 0, 1, 0.7]),
-    )
-    builder.add_box_visual(
-        pose=sapien.Pose(
-            p=[
-                0.03 - grasp_pose_visual_width * 3,
-                -grasp_width - grasp_pose_visual_width,
-                0.03 - 0.05,
-            ],
-            q=quaternions.axangle2quat(np.array([0, 1, 0]), theta=np.pi / 2),
-        ),
-        half_size=[0.04, grasp_pose_visual_width, grasp_pose_visual_width],
-        material=sapien.render.RenderMaterial(base_color=[1, 0, 0, 0.7]),
-    )
-    grasp_pose_visual = builder.build_kinematic(name="grasp_pose_visual")
-    return grasp_pose_visual
-
-def build_panda_gripper_grasp_pose_visual( scene: ManiSkillScene):
-    builder = scene.create_actor_builder()
-    grasp_pose_visual_width = 0.01
-    grasp_width = 0.05
-
-    builder.add_sphere_visual(
-        pose=sapien.Pose(p=[0, 0, 0.0]),
-        radius=grasp_pose_visual_width,
-        material=sapien.render.RenderMaterial(base_color=[0.3, 0.4, 0.8, 0.7])
-    )
-
-    builder.add_box_visual(
-        pose=sapien.Pose(p=[0, 0, -0.08]),
-        half_size=[grasp_pose_visual_width, grasp_pose_visual_width, 0.02],
-        material=sapien.render.RenderMaterial(base_color=[0, 1, 0, 0.7]),
-    )
-    builder.add_box_visual(
-        pose=sapien.Pose(p=[0, 0, -0.05]),
-        half_size=[grasp_pose_visual_width, grasp_width, grasp_pose_visual_width],
-        material=sapien.render.RenderMaterial(base_color=[0, 1, 0, 0.7]),
-    )
-    builder.add_box_visual(
-        pose=sapien.Pose(
-            p=[
-                0.03 - grasp_pose_visual_width * 3,
-                grasp_width + grasp_pose_visual_width,
-                0.03 - 0.05,
-            ],
-            q=quaternions.axangle2quat(np.array([0, 1, 0]), theta=np.pi / 2),
-        ),
-        half_size=[0.04, grasp_pose_visual_width, grasp_pose_visual_width],
-        material=sapien.render.RenderMaterial(base_color=[0, 0, 1, 0.7]),
-    )
-    builder.add_box_visual(
-        pose=sapien.Pose(
-            p=[
-                0.03 - grasp_pose_visual_width * 3,
-                -grasp_width - grasp_pose_visual_width,
-                0.03 - 0.05,
-            ],
-            q=quaternions.axangle2quat(np.array([0, 1, 0]), theta=np.pi / 2),
-        ),
-        half_size=[0.04, grasp_pose_visual_width, grasp_pose_visual_width],
-        material=sapien.render.RenderMaterial(base_color=[1, 0, 0, 0.7]),
-    )
-    grasp_pose_visual = builder.build_kinematic(name="grasp_pose_visual")
-    return grasp_pose_visual
-
-def build_so100_gripper_grasp_pose_visual(scene: ManiSkillScene):
-    builder = scene.create_actor_builder()
-    grasp_pose_visual_width = 0.01
-    grasp_width = 0.05
-
-    builder.add_sphere_visual(
-        pose=sapien.Pose(p=[0, 0, 0.0]),
-        radius=grasp_pose_visual_width,
-        material=sapien.render.RenderMaterial(base_color=[0.3, 0.4, 0.8, 0.7])
-    )
-
-    builder.add_box_visual(
-        pose=sapien.Pose(p=[0, 0, -0.08]),
-        half_size=[grasp_pose_visual_width, grasp_pose_visual_width, 0.02],
-        material=sapien.render.RenderMaterial(base_color=[0, 1, 0, 0.7]),
-    )
-    builder.add_box_visual(
-        pose=sapien.Pose(p=[0, 0, -0.05]),
-        half_size=[grasp_pose_visual_width, grasp_width, grasp_pose_visual_width],
-        material=sapien.render.RenderMaterial(base_color=[0, 1, 0, 0.7]),
-    )
-    builder.add_box_visual(
-        pose=sapien.Pose(
-            p=[
-                0.03 - grasp_pose_visual_width * 3,
-                grasp_width + grasp_pose_visual_width,
-                0.03 - 0.05,
-            ],
-            q=quaternions.axangle2quat(np.array([0, 1, 0]), theta=np.pi / 2),
-        ),
-        half_size=[0.04, grasp_pose_visual_width, grasp_pose_visual_width],
-        material=sapien.render.RenderMaterial(base_color=[0, 0, 1, 0.7]),
-    )
-    builder.add_box_visual(
-        pose=sapien.Pose(
-            p=[
-                0.03 - grasp_pose_visual_width * 3,
-                -grasp_width - grasp_pose_visual_width,
-                0.03 - 0.05,
-            ],
-            q=quaternions.axangle2quat(np.array([0, 1, 0]), theta=np.pi / 2),
-        ),
-        half_size=[0.04, grasp_pose_visual_width, grasp_pose_visual_width],
-        material=sapien.render.RenderMaterial(base_color=[1, 0, 0, 0.7]),
-    )
-    grasp_pose_visual = builder.build_kinematic(name="grasp_pose_visual")
-    return grasp_pose_visual
