+++ mani_skill/examples/motionplanning/panda/motionplanner_stick.py
@@ -1,11 +1,8 @@
 import mplib
 import numpy as np
 import sapien
-import trimesh
-from mani_skill.agents.base_agent import BaseAgent
+
 from mani_skill.envs.sapien_env import BaseEnv
-from mani_skill.envs.scene import ManiSkillScene
-from mani_skill.utils.structs.pose import to_sapien_pose
 from mani_skill.examples.motionplanning.base_motionplanner.motionplanner import BaseMotionPlanningSolver
 
 
@@ -23,27 +20,8 @@ class PandaStickMotionPlanningSolver(BaseMotionPlanningSolver):
         joint_vel_limits=0.9,
         joint_acc_limits=0.9,
     ):
-        super().__init__(env, debug, vis, base_pose, visualize_target_grasp_pose, print_env_info, joint_vel_limits, joint_acc_limits, num_links=self.MOVE_GROUP_LINKS)
+        super().__init__(env, debug, vis, base_pose, visualize_target_grasp_pose, print_env_info, joint_vel_limits, joint_acc_limits)
     
-    def follow_path(self, result, refine_steps: int = 0):
-        n_step = result["position"].shape[0]
-        for i in range(n_step + refine_steps):
-            qpos = result["position"][min(i, n_step - 1)]
-            if self.control_mode == "pd_joint_pos_vel":
-                qvel = result["velocity"][min(i, n_step - 1)]
-                action = np.hstack([qpos, qvel])
-            else:
-                action = np.hstack([qpos])
-            obs, reward, terminated, truncated, info = self.env.step(action)
-            self.elapsed_steps += 1
-            if self.print_env_info:
-                print(
-                    f"[{self.elapsed_steps:3}] Env Output: reward={reward} info={info}"
-                )
-            if self.vis:
-                self.base_env.render_human()
-        return obs, reward, terminated, truncated, info
-
     def setup_planner(self):
         link_names = [link.get_name() for link in self.robot.get_links()]
         joint_names = [joint.get_name() for joint in self.robot.get_active_joints()]
@@ -53,8 +31,8 @@ class PandaStickMotionPlanningSolver(BaseMotionPlanningSolver):
             user_link_names=link_names,
             user_joint_names=joint_names,
             move_group="panda_hand_tcp",
-            joint_vel_limits=np.ones(self.num_links) * self.joint_vel_limits,
-            joint_acc_limits=np.ones(self.num_links) * self.joint_acc_limits,
+            joint_vel_limits=np.ones(self.MOVE_GROUP_LINKS) * self.joint_vel_limits,
+            joint_acc_limits=np.ones(self.MOVE_GROUP_LINKS) * self.joint_acc_limits,
         )
         planner.set_base_pose(np.hstack([self.base_pose.p, self.base_pose.q]))
         return planner
