+++ mani_skill/examples/motionplanning/so100/motionplanner.py
@@ -1,18 +1,14 @@
 import mplib
 import numpy as np
 import sapien
-import trimesh
 
-from mani_skill.agents.base_agent import BaseAgent
 from mani_skill.envs.sapien_env import BaseEnv
-from mani_skill.envs.scene import ManiSkillScene
 from mani_skill.utils.structs.pose import to_sapien_pose
-from transforms3d import quaternions, euler
-from mani_skill.examples.motionplanning.base_motionplanner.motionplanner import BaseMotionPlanningSolver
-from mani_skill.examples.motionplanning.base_motionplanner.gripper_pose_visual_builders import build_so100_gripper_grasp_pose_visual
+from transforms3d import euler
+from mani_skill.examples.motionplanning.two_finger_gripper.motionplanner import TwoFingerGripperMotionPlanningSolver
 
 
-class SO100ArmMotionPlanningSolver (BaseMotionPlanningSolver):
+class SO100ArmMotionPlanningSolver (TwoFingerGripperMotionPlanningSolver):
     OPEN = 0
     CLOSED = -0.8
     MOVE_GROUP_LINKS = 5
@@ -28,15 +24,7 @@ class SO100ArmMotionPlanningSolver (BaseMotionPlanningSolver):
         joint_vel_limits=0.9,
         joint_acc_limits=0.9,
     ):
-        super().__init__(env, debug, vis, base_pose, visualize_target_grasp_pose, print_env_info, joint_vel_limits, joint_acc_limits, num_links=self.MOVE_GROUP_LINKS)
-        if self.vis and self.visualize_target_grasp_pose:
-            if "grasp_pose_visual" not in self.base_env.scene.actors:
-                self.grasp_pose_visual = build_so100_gripper_grasp_pose_visual(
-                    self.base_env.scene
-                )
-            else:
-                self.grasp_pose_visual = self.base_env.scene.actors["grasp_pose_visual"]
-            self.grasp_pose_visual.set_pose(self.base_env.agent.tcp_pose)
+        super().__init__(env, debug, vis, base_pose, visualize_target_grasp_pose, print_env_info, joint_vel_limits, joint_acc_limits)
         self._so_100_visual_grasp_pose_transform = sapien.Pose(q=euler.euler2quat(0, -np.pi / 2, np.pi / 2))
 
     @property
@@ -52,8 +40,8 @@ class SO100ArmMotionPlanningSolver (BaseMotionPlanningSolver):
             user_link_names=link_names,
             user_joint_names=joint_names,
             move_group="Fixed_Jaw_tip",
-            joint_vel_limits=np.ones(self.num_links) * self.joint_vel_limits,
-            joint_acc_limits=np.ones(self.num_links) * self.joint_acc_limits,
+            joint_vel_limits=np.ones(self.MOVE_GROUP_LINKS) * self.joint_vel_limits,
+            joint_acc_limits=np.ones(self.MOVE_GROUP_LINKS) * self.joint_acc_limits,
         )
         planner.set_base_pose(np.hstack([self.base_pose.p, self.base_pose.q]))
         return planner
