+++ mani_skill/examples/teleoperation/interactive_panda.py
@@ -45,6 +45,7 @@ def main(args: Args):
         control_mode="pd_joint_pos",
         render_mode="rgb_array",
         reward_mode="none",
+        robot_uids=args.robot_uid,
         enable_shadow=True,
         viewer_camera_configs=dict(shader_pack=args.viewer_shader)
     )
@@ -90,6 +91,7 @@ def main(args: Args):
             control_mode="pd_joint_pos",
             render_mode="rgb_array",
             reward_mode="none",
+            robot_uids=args.robot_uid,
             human_render_camera_configs=dict(shader_pack=args.video_saving_shader),
         )
         env = RecordEpisode(
